+++ src/com/intellij/idea/plugin/hybris/impex/actions/ImpExValidateAction.kt
@@ -20,12 +20,17 @@ package com.intellij.idea.plugin.hybris.impex.actions
 import com.intellij.idea.plugin.hybris.actions.ExecuteStatementAction
 import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.impex.ImpexLanguage
+import com.intellij.idea.plugin.hybris.impex.editor.impexSplitEditor
 import com.intellij.idea.plugin.hybris.tools.remote.console.impl.HybrisImpexConsole
 import com.intellij.idea.plugin.hybris.tools.remote.execution.impex.ExecutionMode
 import com.intellij.idea.plugin.hybris.tools.remote.execution.impex.ImpExExecutionClient
 import com.intellij.idea.plugin.hybris.tools.remote.execution.impex.ImpExExecutionContext
 import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.application.readAction
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.util.Key
+import com.intellij.ui.AnimatedIcon
+import kotlinx.coroutines.launch
 
 class ImpExValidateAction : ExecuteStatementAction<HybrisImpexConsole>(
     ImpexLanguage,
@@ -34,17 +39,49 @@ class ImpExValidateAction : ExecuteStatementAction<HybrisImpexConsole>(
     "Validate ImpEx file via remote SAP Commerce instance",
     HybrisIcons.ImpEx.Actions.VALIDATE
 ) {
+    override fun update(e: AnActionEvent) {
+        super.update(e)
+
+        val queryExecuting = e.impexSplitEditor()
+            ?.getUserData(KEY_QUERY_EXECUTING)
+            ?: false
+
+        e.presentation.isEnabledAndVisible = e.presentation.isEnabledAndVisible
+        e.presentation.isEnabled = e.presentation.isEnabledAndVisible && !queryExecuting
+        e.presentation.disabledIcon = if (queryExecuting) AnimatedIcon.Default.INSTANCE
+        else HybrisIcons.ImpEx.Actions.VALIDATE
+    }
+
     override fun actionPerformed(e: AnActionEvent, project: Project, content: String) {
-        val console = openConsole(project, content) ?: return
-        val project = e.project ?: return
+        val fileEditor = e.impexSplitEditor()
         val context = ImpExExecutionContext(
             content = content,
             executionMode = ExecutionMode.VALIDATE
         )
 
-        ImpExExecutionClient.getInstance(project).execute(context) { coroutineScope, result ->
-            console.print(result)
+        if (fileEditor?.inEditorResults ?: false) {
+            fileEditor.putUserData(KEY_QUERY_EXECUTING, true)
+            fileEditor.showLoader()
+
+            ImpExExecutionClient.getInstance(project).execute(context) { coroutineScope, result ->
+                fileEditor.renderExecutionResult(result)
+                fileEditor.putUserData(KEY_QUERY_EXECUTING, false)
+
+                coroutineScope.launch {
+                    readAction { this@ImpExValidateAction.update(e) }
+                }
+            }
+        } else {
+            val console = openConsole(project, content) ?: return
+
+            ImpExExecutionClient.getInstance(project).execute(context) { coroutineScope, result ->
+                console.print(result)
+            }
         }
+    }
+
+    companion object {
+        private val KEY_QUERY_EXECUTING = Key.create<Boolean>("impex.query.validate.state")
     }
 
 }